requirements.txt
#Utils
requests==2.31.0
Unidecode==1.3.8

#OCR
paddleocr==2.7.3
paddlepaddle==2.6.1

#MP4
moviepy==1.0.3
SpeechRecognition==3.8.1

# Indexação
elasticsearch==7.15.1

# ChatGPT
openai==1.10.0

# Gemini
#google-ai-generativelanguage==0.4.0
google-generativeai==0.5.2
